requirements.txt
annotated-types==0.7.0
anyio==4.8.0
azure-core==1.33.0
azure-storage-blob==12.25.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
fastapi==0.109.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
isodate==0.7.2
numpy==1.26.2
packaging==24.2
pandas==2.1.4
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
PyMySQL==1.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.20
pytz==2024.2
requests==2.31.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.35.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.27.0
